requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.24
pillow>=10.0
scikit-learn>=1.3
matplotlib>=3.7
fastapi>=0.110
uvicorn>=0.23
python-multipart>=0.0.9
opencv-python>=4.8
